Fix a bug with list menu offsets after changes
NOT sure if this commit is really needed.

if you deposit, toss, withdraw, etc. something from your list menu, you can end
up with cancel on the 2nd or 1st row. this causes a bug where you can scroll below
the cancel option in the list. just adapt the scroll offset in the cases it's too
close to the bottom.